+++ src/lib/features/FeatureTypes.ts
@@ -1,16 +1,5 @@
 import { FeatureValue } from "./FeatureFlags";
 
-export interface FeatureMetadata {
-  name: string;
-  description: string;
-  defaultValue: boolean;
-}
-
-export interface FeatureInfo {
-  key: string;
-  metadata: FeatureMetadata;
-}
-
 export interface FeaturesDoc {
   currentVersion: string;
   features: Record<
